split the orignal analysis script to eda and analysis
