+++ Makefile
@@ -5,12 +5,9 @@ scripts/output.sql:
 		python scripts/to_ddl_script.py
 
 images/year_percentage_lost.png images/year_temperature.png images/year_pesticide.png images/year_aqi.png images/loss_disease_parasite.png images/population_change.png images/temp_change.png:
-		python scripts/analysis_script.py
+		python scripts/eda_script.py
 
-images/average_temperature_linearity.png images/co_conc_linearity.png images/no2_conc_linearity.png images/so2_conc_linearity.png images/percent_lost_by_disease_linearity.png images/pesticide_estimate_linearity.png:
-		python scripts/analysis_script.py
-
-images/correlation_matrix.png models/linear_model.pkl images/shap_train.png images/shap_overall.png data/processed/vif.csv:
+images/average_temperature_linearity.png images/co_conc_linearity.png images/no2_conc_linearity.png images/so2_conc_linearity.png images/percent_lost_by_disease_linearity.png images/pesticide_estimate_linearity.png images/correlation_matrix.png models/linear_model.pkl images/shap_train.png images/shap_overall.png data/processed/vif.csv:
 		python scripts/analysis_script.py
 
 docs/index.html:
@@ -19,7 +16,6 @@ docs/index.html:
 clean:
 		rm -rf scripts/output.sql
 		rm -rf models/linear_model.pkl 
-		rm -rf data/processed/vif.csv
 		rm -rf images/*
 		rm -rf data/processed/*
 		rm -rf docs/*
