+++ docs/reports/analysis_report.html
@@ -410,7 +410,7 @@ Table&nbsp;2: Results of linear regression.
 =========================================================================
 Model:                 OLS                Adj. R-squared:       0.110    
 Dependent Variable:    PercentLost        AIC:                  1058.9125
-Date:                  2024-06-07 03:15   BIC:                  1087.3956
+Date:                  2024-06-07 07:08   BIC:                  1087.3956
 No. Observations:      175                Log-Likelihood:       -520.46  
 Df Model:              8                  F-statistic:          3.679    
 Df Residuals:          166                Prob (F-statistic):   0.000561
